refactor(view): reorganize view package structure by input/output
