+++ src/main/java/lotto/view/input/InputMessage.java
@@ -1,10 +1,10 @@
-package lotto.view.message;
+package lotto.view.input;
 
 public final class InputMessage {
-    private InputMessage() {
-    }
-
     public static final String PURCHASE_AMOUNT = "구입금액을 입력해 주세요.";
     public static final String WINNING_NUMBERS = "당첨 번호를 입력해 주세요.";
     public static final String BONUS_NUMBER = "\n보너스 번호를 입력해 주세요.";
+
+    private InputMessage() {
+    }
 }
