+++ src/main/java/lotto/view/output/OutputMessage.java
@@ -1,4 +1,4 @@
-package lotto.view.message;
+package lotto.view.output;
 
 public final class OutputMessage {
     public static final String LOTTO_COUNT_OUTPUT_FORMAT = "\n%d개를 구매했습니다.\n";
